fix(sizeDiff): unique fail message for size diffs
Previously if snapshots were differently sized, an error would be thrown with the same message as if the image diff had passed the configured threshold. This gave the impression that jesst-image-snapshot was ignoring the failureThreshold config when differently sized images would fail even if the pixel or percentage difference between the images was below the configured threshold.