fix: issue #210
Issue results from max buffer size being fixed to 10MB.  Replaced with
maximum allowable memory so that it becomes the imperative of the user
if the user performs tasks that cause an out of memory error.  This
behavior is then consistent with Node.JS, if they were to allocate
an extraordinarily large Buffer exceeding the available memory in
the system.